Moved the result processor into runtime package
