+++ runtime/processor.go
@@ -1,4 +1,4 @@
-package result
+package runtime
 
 import (
 	"fmt"
